Change data type, so as to avoid casting from int64 to int32

+++ src/impute_by_sample_matching.py
@@ -145,7 +145,7 @@ def prepare_input_matrix(ref_ts, target_ds):
     H1 = np.full(
         (num_sites, num_samples),
         tskit.MISSING_DATA,
-        dtype=np.int64
+        dtype=np.int32
     )
 
     i = 0
